#60 code review
