+++ main.py
@@ -1,7 +1,7 @@
 from Computer_Vision.inference_pipeline import *
 from Communication.serial_functions.functions import *
-from init_stop_sequences.init_stop_sequences import *
-
+from init_stop.init_stop_sequences import *
+#from init_stop.init_stop_sequences_test import *
 
 #encoding = "utf-8"
 #baudrate = 115200
@@ -14,7 +14,9 @@ if __name__ == '__main__':
         #cam, port, ser = init_fake_sequence()
         
         #while True:
-        send_inference_result(cam, model_path)
+        nut = get_inference_nut(cam, model_path)
+        if nut[2] != -1:
+            print_sent_data(send_data(ser, nut_to_string(x, y, nut_class)))
         data = wait_for_data(ser, "")
         #endWhile
